requirements.txt
pytesseract
Pillow
openai
sympy
matplotlib
python-dotenv